* Some methods of MyDict have been moved to functions in another module. * Fixed the "get" method. The default value wasn't been used when the path doesn't exist. * Added missing test for the mentioned fix.
